abstract app component
