+++ lib/components/index.js
@@ -1,31 +1,7 @@
 import 'babel-polyfill';
-import React, { Component } from 'react';
+import React from 'react';
 import { render } from 'react-dom';
-
-export default class App extends Component {
-  state = {
-    test: 'answer'
-  }
-
-  asyncFunc = () => {
-    return Promise.resolve('async response');
-  }
-
-  async componentDidMount() {
-    this.setState({
-      test: await this.asyncFunc(),
-    });
-  }
-
-  render() {
-    return (
-      <div>
-        {this.state.test}
-      </div>
-    );
-  }
-}
-
+import App from './App';
 
 render(
   <App />,
